Issue-1191: moved to folder cause too much module members
