+++ wemake_python_styleguide/logic/naming/logical/name_check.py
@@ -1,6 +1,5 @@
 from typing import Iterable
 
-from wemake_python_styleguide import constants
 from wemake_python_styleguide.logic.naming import access
 
 
@@ -140,97 +139,3 @@ def is_too_long_name(
 
     """
     return len(name) > max_length
-
-
-def does_contain_underscored_number(name: str) -> bool:
-    """
-    Checks for names with underscored number.
-
-    >>> does_contain_underscored_number('star_wars_episode2')
-    False
-
-    >>> does_contain_underscored_number('come2_me')
-    False
-
-    >>> does_contain_underscored_number('_')
-    False
-
-    >>> does_contain_underscored_number('z1')
-    False
-
-    >>> does_contain_underscored_number('iso123_456')
-    False
-
-    >>> does_contain_underscored_number('star_wars_episode_2')
-    True
-
-    >>> does_contain_underscored_number('come_2_me')
-    True
-
-    >>> does_contain_underscored_number('come_44_me')
-    True
-
-    >>> does_contain_underscored_number('iso_123_456')
-    True
-
-    """
-    return constants.UNDERSCORED_NUMBER_PATTERN.match(name) is not None
-
-
-def does_contain_consecutive_underscores(name: str) -> bool:
-    """
-    Checks if name contains consecutive underscores in middle of name.
-
-    >>> does_contain_consecutive_underscores('name')
-    False
-
-    >>> does_contain_consecutive_underscores('__magic__')
-    False
-
-    >>> does_contain_consecutive_underscores('__private')
-    False
-
-    >>> does_contain_consecutive_underscores('name')
-    False
-
-    >>> does_contain_consecutive_underscores('some__value')
-    True
-
-    >>> does_contain_consecutive_underscores('__some__value__')
-    True
-
-    >>> does_contain_consecutive_underscores('__private__value')
-    True
-
-    >>> does_contain_consecutive_underscores('some_value__')
-    True
-
-    """
-    if access.is_magic(name) or access.is_private(name):
-        return '__' in name.strip('_')
-    return '__' in name
-
-
-def does_contain_unicode(name: str) -> bool:
-    """
-    Check if name contains unicode characters.
-
-    >>> does_contain_unicode('hello_world1')
-    False
-
-    >>> does_contain_unicode('')
-    False
-
-    >>> does_contain_unicode('привет_мир1')
-    True
-
-    >>> does_contain_unicode('russian_техт')
-    True
-
-    """
-    try:
-        name.encode('ascii')
-    except UnicodeEncodeError:
-        return True
-    else:
-        return False
